+++ client/src/test/java/org/realityforge/gwt/appcache/client/TestApplicationCache.java
@@ -19,6 +19,11 @@ final class TestApplicationCache
   }
 
   @Override
+  public void abort()
+  {
+  }
+
+  @Override
   public boolean swapCache()
   {
     return true;